requirements.txt
pydantic>=2.7,<3.0
PyYAML>=6.0
requests>=2.31.0
typer>=0.12.3
rich>=13.7.1
pytest>=8.2.0
